Refactor and style action button and success toast.
